Update README.md
The findFloor method finds the floor index of a given integer x in a sorted integer array arr.

The floor of x in a sorted array is the largest element in the array that is less than or equal to x. The method returns the index of this floor element. If no such element exists (i.e., all elements are greater than x), it returns -1.